+++ bson/code.py
@@ -16,7 +16,7 @@
 """
 import collections
 
-from bson.py3compat import string_type
+from bson.py3compat import string_type, PY3, text_type
 
 
 class Code(str):
@@ -32,12 +32,19 @@ class Code(str):
     the `scope` dictionary.
 
     :Parameters:
-      - `code`: string containing JavaScript code to be evaluated
+      - `code`: A string containing JavaScript code to be evaluated or another
+        instance of Code. In the latter case, the scope of `code` becomes this
+        Code's :attr:`scope`.
       - `scope` (optional): dictionary representing the scope in which
         `code` should be evaluated - a mapping from identifiers (as
-        strings) to values
+        strings) to values. Defaults to ``None``. This is applied after any
+        scope associated with a given `code` above.
       - `**kwargs` (optional): scope variables can also be passed as
-        keyword arguments
+        keyword arguments. These are applied after `scope` and `code`.
+
+    ..versionchanged:: 3.4
+      The default value for :attr:`scope` is ``None`` instead of ``{}``.
+
     """
 
     _type_marker = 13
@@ -47,19 +54,29 @@ class Code(str):
             raise TypeError("code must be an "
                             "instance of %s" % (string_type.__name__))
 
-        self = str.__new__(cls, code)
+        if not PY3 and isinstance(code, text_type):
+            self = str.__new__(cls, code.encode('utf8'))
+        else:
+            self = str.__new__(cls, code)
 
         try:
             self.__scope = code.scope
         except AttributeError:
-            self.__scope = {}
+            self.__scope = None
 
         if scope is not None:
             if not isinstance(scope, collections.Mapping):
                 raise TypeError("scope must be an instance of dict")
-            self.__scope.update(scope)
+            if self.__scope is not None:
+                self.__scope.update(scope)
+            else:
+                self.__scope = scope
 
-        self.__scope.update(kwargs)
+        if kwargs:
+            if self.__scope is not None:
+                self.__scope.update(kwargs)
+            else:
+                self.__scope = kwargs
 
         return self